refactor: move `Router` and `RouterBuilder` into `protocol.rs`
