+++ iroh/src/lib.rs
@@ -242,7 +242,6 @@ pub mod endpoint;
 mod magicsock;
 pub mod metrics;
 pub mod protocol;
-pub mod router;
 pub mod ticket;
 pub mod tls;
